python/S01E05/requirements.txt
requests
PyYAML
numexpr
openai
